requirements.txt
streamlit
ultralytics
torch
torchvision
Pillow
opencv-python
